test: move tests to right class
