+++ armadillo/src/test/java/org/molgenis/armadillo/service/ManagementServiceTest.java
@@ -438,73 +438,6 @@ class ManagementServiceTest {
   }
 
   // -------------------------------------------------------------------------
-  // createPythonScript (private — tested via reflection)
-  // -------------------------------------------------------------------------
-
-  @Test
-  void createPythonScript_nonUpdateBranch_doesNotContainUpdateFlag() throws Exception {
-    Method m =
-        ManagementService.class.getDeclaredMethod(
-            "createPythonScript", String.class, String.class, String.class, Boolean.class);
-    m.setAccessible(true);
-
-    String script =
-        (String)
-            m.invoke(
-                service,
-                "/usr/share/armadillo/armadillo-reboot.sh",
-                "/tmp/update.log",
-                "5.14.0",
-                false);
-
-    assertTrue(script.contains("import os, sys, subprocess"));
-    assertTrue(script.contains("/tmp/update.log"));
-    assertTrue(script.contains("5.14.0"));
-    // The '-u' update flag must NOT be present when isUpdate=false
-    assertFalse(script.contains("'-u'"));
-  }
-
-  @Test
-  void createPythonScript_updateBranch_containsUpdateFlag() throws Exception {
-    Method m =
-        ManagementService.class.getDeclaredMethod(
-            "createPythonScript", String.class, String.class, String.class, Boolean.class);
-    m.setAccessible(true);
-
-    String script =
-        (String)
-            m.invoke(
-                service,
-                "/usr/share/armadillo/armadillo-reboot.sh",
-                "/tmp/update.log",
-                "5.14.0",
-                true);
-
-    assertTrue(script.contains("'-u'"));
-  }
-
-  @Test
-  void createPythonScript_containsArmadilloHomeVersionAndMode() throws Exception {
-    Method m =
-        ManagementService.class.getDeclaredMethod(
-            "createPythonScript", String.class, String.class, String.class, Boolean.class);
-    m.setAccessible(true);
-
-    String script =
-        (String)
-            m.invoke(
-                service,
-                "/usr/share/armadillo/armadillo-reboot.sh",
-                "/tmp/update.log",
-                "5.14.0",
-                false);
-
-    assertTrue(script.contains(tempDir.toString())); // armadilloHome injected via -p
-    assertTrue(script.contains("5.14.0")); // version injected via -v
-    assertTrue(script.contains("PROD")); // mode injected via -m
-  }
-
-  // -------------------------------------------------------------------------
   // isArmadilloUpdateAvailable — jar already present (no HTTP call needed)
   // -------------------------------------------------------------------------
